test swap

+++ app/src/main/java/org/p2p/wallet/swap/jupiter/api/request/CreateSwapTransactionRequest.kt
@@ -62,7 +62,7 @@ data class SwapRouteRequest(
         @SerializedName("minOutAmount")
         val minOutAmount: String?,
         @SerializedName("priceImpactPct")
-        val priceImpactPct: Int,
+        val priceImpactPct: Double,
         @SerializedName("lpFee")
         val lpFee: LpFeeRequest,
         @SerializedName("platformFee")
@@ -74,7 +74,7 @@ data class SwapRouteRequest(
             @SerializedName("mint")
             val mint: String,
             @SerializedName("pct")
-            val pct: Int
+            val pct: Double
         )
 
         data class PlatformFeeRequest(
@@ -83,7 +83,7 @@ data class SwapRouteRequest(
             @SerializedName("mint")
             val mint: String,
             @SerializedName("pct")
-            val pct: Int
+            val pct: Double
         )
     }
 }
